requirements.txt
groq==1.0.0
python-dotenv==1.2.1
rich==15.0.0
